Initial clean up from lecture
