+++ SecretSanta/src/SecretSanta.Web/appsettings.Development.json
@@ -7,6 +7,6 @@
     }
   },
   "SecretSantaApi": {
-    "BaseUri": "https://localhost:44393/"
+    "BaseUri": "https://localhost:44310/"
   }
 }
