+++ SecretSanta/test/SecretSanta.Web.UITests/Pages/AddGroupsPage.cs
@@ -1,0 +1,28 @@
+﻿using OpenQA.Selenium;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+namespace SecretSanta.Web.UITests
+{
+    public class AddGroupsPage
+    {
+
+        public const string Slug = GroupsPage.Slug + "/Add";
+
+        public IWebDriver Driver { get; }
+
+        public IWebElement GroupNameTextBox => Driver.FindElement(By.Id("Name"));
+
+        public IWebElement SubmitButton =>
+            Driver
+                .FindElements(By.CssSelector("button.is-primary"))
+                .Single(x => x.Text == "Submit");
+
+        public AddGroupsPage(IWebDriver driver)
+        {
+            Driver = driver ?? throw new ArgumentNullException(nameof(driver));
+        }
+    }
+}
